% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_raster_to_county.R
\name{clip_raster_to_county}
\alias{clip_raster_to_county}
\title{Clip a raster to a county in Washington State}
\usage{
clip_raster_to_county(path_to_tif, county_name)
}
\arguments{
\item{path_to_tif}{File path to a GeoTIFF raster}

\item{county_name}{Name of the Washington State county (e.g., "Columbia")}
}
\value{
A \code{SpatRaster} object cropped and masked to the county boundary
' @example
\dontrun{


r_mask_2024 <- clip_raster_to_county("../../../Downloads/2024_30m_cdls/2024_30m_cdls.tif",
county_name = "Columbia")

terra::writeRaster(
r_mask_2024,
filename = "cropland_data_layer_2024.tif",
overwrite = TRUE,
filetype = "GTiff",
gdal = c("COMPRESS=LZW")
)



}
}
\description{
\link{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
