% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_esri_dates.R
\name{convert_esri_dates}
\alias{convert_esri_dates}
\title{Convert ESRI Epoch Millisecond Dates to POSIXct}
\usage{
convert_esri_dates(x)
}
\arguments{
\item{x}{A numeric vector of ESRI date values in milliseconds.}
}
\value{
A \code{POSIXct} vector representing human-readable date-times.
}
\description{
Converts numeric ESRI date values (milliseconds since 1970-01-01)
to standard R \code{POSIXct} date-time objects.
}
\examples{
\dontrun{
# LastSurveyDate looks like 1.0508e+12
convert_esri_dates(LastSurveyDate)
}
}
