% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cd_polygon.R
\name{get_cd_polygon}
\alias{get_cd_polygon}
\title{Get Conservation District Boundary Geometry for a Washington County}
\usage{
get_cd_polygon(district_name, crs = 4326)
}
\arguments{
\item{district_name}{Character. Name of the district (e.g., "Columbia" or c("Columbia","Palouse")}

\item{crs}{Integer or character. Desired coordinate reference system (default: 4326)}
}
\value{
An sf object with the swcd boundary
}
\description{
Get Conservation District Boundary Geometry for a Washington County
}
\examples{

get_cd_polygon(c("Columbia","Palouse"))

get_cd_polygon("Columbia")

}
