% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county_polygon.R
\name{get_county_polygon}
\alias{get_county_polygon}
\title{Get County Boundary Geometry for a Washington County}
\usage{
get_county_polygon(county_name, crs = 4326)
}
\arguments{
\item{county_name}{Character. Name of the county (e.g., "Columbia" or c("Columbia","Asotin)}

\item{crs}{Integer or character. Desired coordinate reference system (default: 4326)}
}
\value{
An sf object with the county boundary
}
\description{
Get County Boundary Geometry for a Washington County
}
