% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huc_by_county.R
\name{get_huc_by_county}
\alias{get_huc_by_county}
\title{Load Washington HUC Watersheds by County}
\usage{
get_huc_by_county(
  county_name,
  huc_level = "huc12",
  save_shp = FALSE,
  shp_path = NULL
)
}
\arguments{
\item{county_name}{Character. Name of the county (e.g., "Columbia").}

\item{huc_level}{Character. One of "huc8", "huc10", "huc12". Default is "huc12".}

\item{save_shp}{Logical. Whether to save the result as a shapefile. Default is FALSE.}

\item{shp_path}{Character. File path to save shapefile (without extension). Default is NULL.}
}
\value{
An \code{sf} object with watershed polygons assigned to the input county.
}
\description{
This function downloads Watershed Boundary Dataset (WBD) HUC polygons
for a given county in Washington and clips HUC polygon to the county
provided. Optionally saves the results as a shapefile. Clipped acres of
each watershed are automatically recalculated and added as \code{Acres_in_huc} variable
}
\examples{
 \dontrun{

 asotin_huc12s <- get_huc_by_county("Columbia",
 huc_level = "huc12",
 save_shp = TRUE,
 shp_path = "columbia_huc12")

  leaflet() \%>\%
  addTiles() \%>\%
  addPolygons(data=asotin_huc12s,label=~name)

}
}
