% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_raster_to_cd.R
\name{reclassify_raster}
\alias{reclassify_raster}
\title{Reclassify CDL raster to 7 generalized NLCD-style land cover groups}
\usage{
reclassify_raster(r)
}
\arguments{
\item{r}{A \code{SpatRaster} object representing CDL land cover codes}
}
\value{
A new \code{SpatRaster} with values 1–7 representing generalized land cover categories
and a category table for labeling and mapping.

\link{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}

' @example
\dontrun{

category_colors <- c(
  "Row Crops"        = "#FFFF64",  # Cropland yellow
  "Non-Crop Natural" = "#DCD939",  # Grassland/herbaceous
  "Specialty Crops"  = "#A8E6A0",  # Orchard/vineyard
  "Developed"        = "#FF0000",  # High intensity development
  "Water"            = "#476BA1",  # Open water
  "NLCD Natural"     = "#397D49",  # Evergreen forest
  "Other/No Data"    = "#D3D3D3"   # No data
)

leaflet() %>%
 addTiles() %>%
  leaflet::addRasterImage(r_mask_2024, project = TRUE, opacity = 0.7) %>%
  addLegend(
  position = "bottomright",
  colors = category_colors,
  labels = names(category_colors),
  title = "Land Cover 2011",
  opacity = 1
  )
}
}
\description{
Reclassify CDL raster to 7 generalized NLCD-style land cover groups
}
