% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_geoserver_layers.R
\name{show_geoserver_layers}
\alias{show_geoserver_layers}
\title{Show available layers from a GeoServer instance}
\usage{
show_geoserver_layers(
  geoserver_url =
    "https://geoserver.megaloptera-data.com/geoserver/Columbia/wms?service=WMS&version=1.1.0&request=GetCapabilities"
)
}
\arguments{
\item{geoserver_url}{Character string specifying the base URL of the GeoServer instance}
}
\value{
A character vector of available layer names
}
\description{
This function retrieves and displays the available layers from a GeoServer WFS
(Web Feature Service) capabilities document.
}
\examples{
\dontrun{
# Show available layers
layers <- show_geoserver_layers("http://my-geoserver.com/geoserver")
print(layers)
}
}
