% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swcd_boundaries}
\alias{swcd_boundaries}
\title{Conservation District Boundaries for Washington State}
\format{
An \code{sf} data frame
}
\source{
Washington State GIS Open Data Portal
}
\usage{
swcd_boundaries
}
\description{
A simple features (sf) object containing conservation district boundaries.
}
\keyword{datasets}
