% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wa_counties}
\alias{wa_counties}
\title{Washington State Counties}
\format{
An sf object with 39 rows and the following variables:
\describe{
\item{GEOID}{Geographic identifier for the county}
\item{NAME}{County name}
\item{geometry}{Spatial geometry (polygon boundaries)}
}
}
\source{
US Census Bureau TIGER/Line Shapefiles
}
\usage{
wa_counties
}
\description{
A dataset containing boundary polygons for all counties in Washington State.
This data is derived from the US Census Bureau's TIGER/Line shapefiles.
}
\references{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}
}
\keyword{datasets}
