% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_geoserver_layers.R
\name{show_geoserver_layers}
\alias{show_geoserver_layers}
\title{Get List of Layers from Columbia GeoServer}
\usage{
show_geoserver_layers(
  geoserver_url =
    "http://geoserver.megaloptera-data.com/geoserver/Columbia/wms?service=WMS&version=1.1.0&request=GetCapabilities"
)
}
\arguments{
\item{capabilities_url}{A character string specifying the WMS GetCapabilities URL.
Defaults to the Columbia workspace on your GeoServer:
\code{"http://142.93.92.104:8080/geoserver/Columbia/wms?service=WMS&version=1.1.0&request=GetCapabilities"}}
}
\value{
A character vector of layer names (e.g., \code{"Columbia:usda_2024_cropland"}).
}
\description{
This function queries a GeoServer WMS GetCapabilities endpoint and extracts a list of available WMS layer names.
By default, it points to the Columbia County WA workspace.
}
\examples{
\dontrun{

show_geoserver_layers()
}

}
