% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_acres_by_huc12.R
\name{summarize_acres_by_huc12}
\alias{summarize_acres_by_huc12}
\title{Compute acres of overlay polygons within each HUC12 watershed}
\usage{
summarize_acres_by_huc12(huc12, overlay, huc_id_col = "huc12")
}
\arguments{
\item{huc12}{An \code{sf} object of watershed boundaries (reprojected to 5070)}

\item{overlay}{An \code{sf} object of polygonal features (also in 5070)}

\item{huc_id_col}{Name of the HUC identifier column (default = "huc12")}
}
\value{
A tibble summarizing acres of overlay per HUC12 unit
}
\description{
Compute acres of overlay polygons within each HUC12 watershed
}
